fix(tga): remediate trusty-review REQUEST_CHANGES on PR #1117 (issue #1113)
Finding 1 (med) — SQL orphan doc: add prominent DOCUMENTATION/REFERENCE ONLY
banner to 0021_agentic_mode.sql so readers know the live migration is in v21.rs.

Finding 2 (med) — name→email silent fallback: persist_weekly_engineer now
emits tracing::warn! and SKIPS the row when name_to_email.get(author_display)
misses, preserving grain-key integrity instead of silently corrupting it.

Finding 3 (med) — cursor regex false positives: tighten detect_agentic_mode
pattern to (?i)@cursor\.sh|\bCursor\b plus a code-level is_cursor_match()
helper that rejects matches followed by `-` (hyphenated surname guard, since
Rust's regex crate has no lookahead). Add test for the Alice Cursor-Williams
false-positive case.

Finding 4 (low) — AgenticMode parsing: add impl std::str::FromStr for
AgenticMode; use .parse::<AgenticMode>().unwrap_or(AgenticMode::None) in
aggregator.rs instead of the hand-rolled match. Add round-trip test.

Finding 5 (low) — agentic_pct comment: add inline comment in persist.rs
stating agentic_pct = agentic_count / net * 100 intentionally EXCLUDES
ide_assisted_count (full-agentic only, per #1113 spec).

Finding 6 (line-cap) — revert allowlist budget bumps: extractor.rs reverted
to frozen budget 1318 (compressed two comment blocks), aggregator.rs reverted
to frozen budget 1707 (extracted both persist functions to new report/persist.rs
submodule with thin delegating stubs maintaining the public API). ai_attribution.rs
stays at 499 lines (≤ 500 cap, not allowlisted). check_line_cap.sh passes with
0 violations and no budget increases on tga files.
