+++ .line-cap-allowlist.tsv
@@ -72,7 +72,7 @@ crates/trusty-git-analytics/src/classify/tiers/weighted_sum.rs	962
 crates/trusty-git-analytics/src/collect/azdo/client.rs	2658
 crates/trusty-git-analytics/src/collect/azdo/pr_fetcher.rs	1707
 crates/trusty-git-analytics/src/collect/collector.rs	1181
-crates/trusty-git-analytics/src/collect/git/extractor.rs	1321
+crates/trusty-git-analytics/src/collect/git/extractor.rs	1318
 crates/trusty-git-analytics/src/collect/git/reachability.rs	1432
 crates/trusty-git-analytics/src/collect/github/client.rs	1225
 crates/trusty-git-analytics/src/collect/identity/resolver.rs	781
@@ -84,7 +84,7 @@ crates/trusty-git-analytics/src/commands/incidents.rs	828
 crates/trusty-git-analytics/src/core/config/mod.rs	1624
 crates/trusty-git-analytics/src/core/config/validator.rs	816
 crates/trusty-git-analytics/src/core/effort_percentile.rs	531
-crates/trusty-git-analytics/src/report/aggregator.rs	1850
+crates/trusty-git-analytics/src/report/aggregator.rs	1707
 crates/trusty-git-analytics/src/report/drilldown.rs	1010
 crates/trusty-git-analytics/src/report/mod.rs	916
 crates/trusty-gworkspace/src/tools.rs	598
